requirements.txt
torchvision==0.13.1
transformers==4.22.1
numpy==1.23.1
pandas==1.4.4
torch==1.12.1
Pillow==9.2.0
